requirements.txt
pandas>=2.2
numpy>=1.26
scikit-learn>=1.5
tqdm>=4.66
python-dotenv>=1.0
typer[all]>=0.12
cfbd>=4.5.3
pydantic>=1.10.5,<2.0
